+++ packages/moonstone/CHANGELOG.md
@@ -12,6 +12,8 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Fixed
 
+- `moonstone/Scrollable` to update disability of paging controls when the scrollbar is set to `visible` and the content becomes shorter.
+
 ## [1.5.0] - 2017-07-19
 
 ### Added